+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#1281](https://github.com/alleslabs/celatone-frontend/pull/1281) Add advanced options helper text, limit fetching verification infos to hex address only, add more module information
 - [#1279](https://github.com/alleslabs/celatone-frontend/pull/1279) Add Initia App Menu from library
 - [#1276](https://github.com/alleslabs/celatone-frontend/pull/1276) Support backward LCD compatibility for custom networks